ci: Clean up Browserstack test files, remove unused files (#788)
