+++ package.json
@@ -37,8 +37,8 @@
         "build:rollup:requirejs": "rollup --config test/integrations/requirejs/rollup.requirejs.config.js",
         "build:ts": "tsc -p .",
         "test": "npm run build && npm run build:test-bundle && cross-env DEBUG=false karma start test/karma.config.js",
-        "test:browserstack": "karma start test/cross-browser-testing/browserstack/browserstack.karma.config.js",
-        "test:browserstack:debug": "cross-env DEBUG=true karma start test/cross-browser-testing/browserstack/browserstack.karma.config.js",
+        "test:browserstack": "karma start test/cross-browser-testing/browserstack.karma.config.js",
+        "test:browserstack:debug": "cross-env DEBUG=true karma start test/cross-browser-testing/browserstack.karma.config.js",
         "test:debug": "cross-env DEBUG=true karma start test/karma.config.js",
         "test:stub": "cross-env TESTTYPE=stub ENVIRONMENT=prod rollup --config rollup.test.config.js && karma start test/stub/karma.stub.config.js",
         "test:integrations": "npm run test:requirejs && npm run test:integrations:cjs && npm run test:integrations:module",
