+++ test/karma.config.js
@@ -1,9 +1,7 @@
 const { DEBUG } = process.env;
 
-let files = [
-    'lib/mockhttprequest.js',
+const files = [
     'lib/geomock.js',
-    '../node_modules/sinon-browser-only/sinon.js',
     'config.js',
     '../dist/mparticle.js',
     'test-bundle.js',
