+++ test/src/tests-main.js
@@ -28,29 +28,29 @@ beforeEach(function() {
 });
 
 import './tests-core-sdk';
-// import './tests-batchUploader';
-// import './tests-beaconUpload';
-// import './tests-kit-blocking';
-// import './tests-persistence';
-// import './tests-forwarders';
-// import './tests-helpers';
-// import './tests-identity';
-// import './tests-event-logging';
-// import './tests-eCommerce';
-// import './tests-cookie-syncing';
-// import './tests-identities-attributes';
-// import './tests-native-sdk';
-// import './tests-consent';
-// import './tests-serverModel';
-// import './tests-mockBatchCreator.ts';
-// import './tests-mParticleUser';
-// import './tests-self-hosting-specific';
-// import './tests-runtimeToBatchEventsDTO';
-// import './tests-apiClient';
-// import './tests-mparticle-instance-manager';
-// import './tests-queue-public-methods';
-// import './tests-validators';
-// import './tests-utils';
-// import './tests-session-manager';
-// import './tests-store';
-// import './tests-config-api-client';
+import './tests-batchUploader';
+import './tests-beaconUpload';
+import './tests-kit-blocking';
+import './tests-persistence';
+import './tests-forwarders';
+import './tests-helpers';
+import './tests-identity';
+import './tests-event-logging';
+import './tests-eCommerce';
+import './tests-cookie-syncing';
+import './tests-identities-attributes';
+import './tests-native-sdk';
+import './tests-consent';
+import './tests-serverModel';
+import './tests-mockBatchCreator.ts';
+import './tests-mParticleUser';
+import './tests-self-hosting-specific';
+import './tests-runtimeToBatchEventsDTO';
+import './tests-apiClient';
+import './tests-mparticle-instance-manager';
+import './tests-queue-public-methods';
+import './tests-validators';
+import './tests-utils';
+import './tests-session-manager';
+import './tests-store';
+import './tests-config-api-client';
